+++ testfiles/simple_usage.py
@@ -202,6 +202,10 @@ async def run() -> None:
         llm = init_chat_model("google_genai:gemini-2.5-flash")
         # llm = init_chat_model("google_genai:gemini-2.5-pro")
 
+        ### https://console.x.ai
+        # llm = init_chat_model("xai:grok-3-mini")
+        # llm = init_chat_model("xai:grok-4")
+
         agent = create_react_agent(
             llm,
             tools
@@ -239,6 +243,7 @@ async def run() -> None:
         print("\x1b[0m")   # reset the color
 
     finally:
+        # cleanup can be undefined when an exeption occurs during initialization
         if "cleanup" in locals():
             await cleanup()